code/scai/pyproject.toml
[project]
name = "scai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
